src/server-python/requirements.txt
# Install the packages from PyPi repository.
SQLAlchemy>=1.4.36
python-dotenv==0.21.0